+++ src/PHPCR/Shell/Console/Command/Phpcr/NodeListCommand.php
@@ -100,7 +100,11 @@ HERE
             $this->renderNode($node, $table, array(), $filter);
 
             if ($table->getNumberOfRows() > 0) {
-                $output->writeln('<path>' . $node->getPath() . '</path>');
+                $output->writeln(sprintf('<path>%s</path> [%s] > %s',
+                    $node->getPath(),
+                    $node->getPrimaryNodeType()->getName(),
+                    implode(', ', $node->getPrimaryNodeType()->getDeclaredSupertypeNames())
+                ));
                 $table->render($output);
             }
         }
@@ -224,7 +228,11 @@ HERE
 
             $table->addRow(array(
                 '<property>' . implode('', $spacers). $name . '</property>',
-                '<property-type>' . $this->formatter->getPropertyTypeName($property->getType()) . '</property-type>',
+                sprintf(
+                    '<property-type>%s (%s)</property-type>',
+                    $this->formatter->getPropertyTypeName($property->getType()),
+                    implode(',', (array) $property->getLength())
+                ),
                 $valueCell,
             ));
         }