+++ CHANGELOG.md
@@ -19,6 +19,8 @@ dev-master
 
 ### Enhancements
 
+- [node:list] Show primary node type and super-types
+- [node:list] Show property length
 - [autocomplete] Autocomplete now includes command names
 - [references] Show UUIDs when listing reference properties
 - [import/export] Renamed session import and export to `session:import` &